+++ tests/functional_tests/expected_values/amazon.py
@@ -6,10 +6,10 @@ conf = fixtures.get_configuration('amazon')
 
 
 CONFIG = {
+    'logout_url': 'http://www.amazon.com/gp/flex/sign-out.html',
     'login_xpath': '//*[@id="ap_email"]',
     'password_xpath': '//*[@id="ap_password"]',
     'consent_xpaths': [
-        '//*[@id="signInSubmit"]/span/button',
         '//*[@id="ap-oaconsent-agree-button"]/span/button',
     ],
     'class_': oauth2.Amazon,
@@ -25,6 +25,7 @@ CONFIG = {
         'last_name': None,
         'link': None,
         'locale': None,
+        'location': None,
         'name': conf.user_name,
         'nickname': None,
         'phone': None,