+++ tests/functional_tests/expected_values/xero.py
@@ -1,28 +1,31 @@
+import re
+
 import fixtures
 import constants
-from authomatic.providers import oauth2
+from authomatic.providers import oauth1
 
-conf = fixtures.get_configuration('eventbrite')
+conf = fixtures.get_configuration('xero')
 
 CONFIG = {
-    'logout_url': 'https://www.eventbrite.com/logout',
-    'login_xpath': '//*[@id="login-email"]',
-    'password_xpath': '//*[@id="login-password"]',
+    'logout_url': 'https://my.xero.com/logout',
+    'login_xpath': '//*[@id="email"]',
+    'password_xpath': '//*[@id="password"]',
     'consent_xpaths': [
-        '//*[@id="access_choices_allow"]',
+        '//*[@id="SelectedOrganisation_toggle"]',
+        '//*[@id="SelectedOrganisation_suggestions"]/div/div',
+        '//*[@id="ext-gen4"]'
     ],
-    'before_login_enter_wait': 2,
-    'consent_wait_seconds': 2,
-    'class_': oauth2.Eventbrite,
-    'scope': oauth2.Eventbrite.user_info_scope,
+    'consent_wait_seconds': 5,
+    'after_consent_wait_seconds': 10,
+    'class_': oauth1.Xero,
     'user': {
         'birth_date': None,
         'city': None,
         'country': None,
         'email': conf.user_email,
-        'first_name': conf.user_first_name,
         'gender': None,
-        'id': conf.user_id,
+        'id': re.compile(r'\w{8}-\w{4}-\w{4}-\w{4}-\w{12}'),
+        'first_name': conf.user_first_name,
         'last_name': conf.user_last_name,
         'link': None,
         'locale': None,
@@ -40,36 +43,35 @@ CONFIG = {
         conf.user_first_name,
         conf.user_last_name,
         conf.user_name,
-        conf.user_id,
 
         # User info JSON keys
-        'emails', 'email', 'verified', 'primary', 'id', 'name', 'first_name',
-        'last_name'
     ],
     # Case insensitive
     'content_should_not_contain':
         conf.no_birth_date +
         conf.no_gender +
+        conf.no_link +
         conf.no_locale +
         conf.no_location +
         conf.no_nickname +
         conf.no_phone +
-        conf.no_timezone
-        ,
+        conf.no_picture +
+        conf.no_timezone +
+        conf.no_username,
     # True means that any thruthy value is expected
     'credentials': {
-        'token_type': 'Bearer',
-        'provider_type_id': '2-17',
         '_expiration_time': None,
-        'consumer_key': None,
-        'provider_id': None,
-        'consumer_secret': None,
-        'token': True,
-        'token_secret': None,
         '_expire_in': True,
-        'provider_name': 'eventbrite',
-        'refresh_token': None,
-        'provider_type': 'authomatic.providers.oauth2.OAuth2',
+        'consumer_key': True,
+        'consumer_secret': True,
+        'provider_id': None,
+        'provider_name': 'xero',
+        'provider_type': 'authomatic.providers.oauth1.OAuth1',
+        'provider_type_id': '1-9',
         'refresh_status': constants.CREDENTIALS_REFRESH_NOT_SUPPORTED,
+        'refresh_token': None,
+        'token': True,
+        'token_secret': True,
+        'token_type': None,
     },
 }